fix: Resolve all compiler warnings (11 files)
Critical fixes:
- report.c: Fix format string bugs (undefined behavior)
- loader.c: Use validate_plugin_name() for security (path traversal prevention)

Code quality fixes:
- scheduler.c: Fix NSEC_PER_SEC macro redefinition + sign comparison
- test_clock_resolution.c: Fix NSEC_PER_SEC macro redefinition
- 6 v2 kernels: Add #undef before CORTEX_ABI_VERSION redefinition
- ica.c: Cast uint32_t to size_t in overflow checks (tautological comparison)

Verification:
- make clean && make all: ✅ Zero warnings
- make tests: ✅ All tests passing (21+ tests)

Changes improve code safety (security validation, undefined behavior fix)
and eliminate compiler noise for cleaner builds.

+++ primitives/kernels/v1/ica@f32/ica.c
@@ -463,7 +463,7 @@ cortex_calibration_result_t cortex_calibrate(
     }
 
     /* Check for overflow: C * sizeof(float) */
-    if (C > SIZE_MAX / sizeof(float)) {
+    if ((size_t)C > SIZE_MAX / sizeof(float)) {
         fprintf(stderr, "[ica] ERROR: Integer overflow in mean allocation (C=%u)\n", C);
         free(X);
         return (cortex_calibration_result_t){NULL, 0, 0};
@@ -594,7 +594,7 @@ cortex_init_result_t cortex_init(const cortex_plugin_config_t *config) {
     state->C = C;
 
     /* Check for overflow: C * sizeof(float) */
-    if (C > SIZE_MAX / sizeof(float)) {
+    if ((size_t)C > SIZE_MAX / sizeof(float)) {
         fprintf(stderr, "[ica] ERROR: Integer overflow in state->mean allocation (C=%u)\n", C);
         free(state);
         return (cortex_init_result_t){NULL, 0, 0, 0};
